@@ -1,14 +1,14 @@
 from datetime import timedelta
-from typing import Optional
 
-from taskiq.cli.scheduler.run import run_scheduler_loop
+from taskiq.cli.scheduler.run import SchedulerLoop
 from taskiq.scheduler.scheduler import TaskiqScheduler
 
 
 async def run_scheduler_task(
     scheduler: TaskiqScheduler,
     run_startup: bool = False,
-    interval: Optional[timedelta] = None,
+    interval: timedelta | None = None,
+    loop_interval: timedelta | None = None,
 ) -> None:
     """
     Run scheduler task.
@@ -18,10 +18,16 @@ async def run_scheduler_task(
 
     :param scheduler: scheduler instance.
     :param run_startup: whether to run startup function or not.
+    :param interval: interval to check for schedule updates.
+    :param loop_interval: interval to check tasks to send.
     """
     for source in scheduler.sources:
         await source.startup()
     if run_startup:
         await scheduler.startup()
     while True:
-        await run_scheduler_loop(scheduler, interval)
+        scheduler_loop = SchedulerLoop(scheduler)
+        await scheduler_loop.run(
+            update_interval=interval,
+            loop_interval=loop_interval,
+        )
