@@ -1,21 +1,16 @@
 import sys
-from collections.abc import Coroutine
-from datetime import datetime
+from collections.abc import Callable, Coroutine
+from datetime import datetime, timedelta
 from types import CoroutineType
 from typing import (
     TYPE_CHECKING,
     Any,
-    Callable,
-    Dict,
     Generic,
-    Optional,
-    Type,
+    ParamSpec,
     TypeVar,
     Union,
     overload,
 )
-
-from typing_extensions import ParamSpec
 
 from taskiq.kicker import AsyncKicker
 from taskiq.scheduler.created_schedule import CreatedSchedule
@@ -52,8 +47,8 @@ class AsyncTaskiqDecoratedTask(Generic[_FuncParams, _ReturnType]):
         broker: "AsyncBroker",
         task_name: str,
         original_func: Callable[_FuncParams, _ReturnType],
-        labels: Dict[str, Any],
-        return_type: Optional[Type[_ReturnType]] = None,
+        labels: dict[str, Any],
+        return_type: type[_ReturnType] | None = None,
     ) -> None:
         self.broker = broker
         self.task_name = task_name
@@ -106,24 +101,21 @@ class AsyncTaskiqDecoratedTask(Generic[_FuncParams, _ReturnType]):
         self: "AsyncTaskiqDecoratedTask[_FuncParams, CoroutineType[Any, Any, _T]]",
         *args: _FuncParams.args,
         **kwargs: _FuncParams.kwargs,
-    ) -> AsyncTaskiqTask[_T]:
-        ...
+    ) -> AsyncTaskiqTask[_T]: ...
 
     @overload
     async def kiq(
         self: "AsyncTaskiqDecoratedTask[_FuncParams, Coroutine[Any, Any, _T]]",
         *args: _FuncParams.args,
         **kwargs: _FuncParams.kwargs,
-    ) -> AsyncTaskiqTask[_T]:
-        ...
+    ) -> AsyncTaskiqTask[_T]: ...
 
     @overload
     async def kiq(
         self: "AsyncTaskiqDecoratedTask[_FuncParams, _ReturnType]",
         *args: _FuncParams.args,
         **kwargs: _FuncParams.kwargs,
-    ) -> AsyncTaskiqTask[_ReturnType]:
-        ...
+    ) -> AsyncTaskiqTask[_ReturnType]: ...
 
     async def kiq(
         self,
@@ -165,6 +157,32 @@ class AsyncTaskiqDecoratedTask(Generic[_FuncParams, _ReturnType]):
         return await self.kicker().schedule_by_cron(
             source,
             cron,
+            *args,
+            **kwargs,
+        )
+
+    async def schedule_by_interval(
+        self,
+        source: "ScheduleSource",
+        interval: int | timedelta,
+        *args: _FuncParams.args,
+        **kwargs: _FuncParams.kwargs,
+    ) -> CreatedSchedule[_ReturnType]:
+        """
+        Schedule the task to start using an interval.
+
+        This method requires a schedule source,
+        which is capable of dynamically adding new schedules.
+
+        :param source: schedule source.
+        :param interval: interval in seconds or timedelta instance.
+        :param args: function's arguments.
+        :param kwargs: function's key word arguments.
+        :return: schedule id.
+        """
+        return await self.kicker().schedule_by_interval(
+            source,
+            interval,
             *args,
             **kwargs,
         )
