@@ -1,8 +1,7 @@
 import enum
-from typing import Generic, Optional, Union
+from typing import Generic, TypeVar
 
 from taskiq_dependencies import Depends
-from typing_extensions import TypeVar
 
 from taskiq.compat import IS_PYDANTIC2
 from taskiq.context import Context
@@ -36,8 +35,8 @@ else:
 class TaskProgress(_TaskProgressConfig, Generic[_ProgressType]):
     """Progress of task execution."""
 
-    state: Union[TaskState, str]
-    meta: Optional[_ProgressType]
+    state: TaskState | str
+    meta: _ProgressType | None
 
 
 class ProgressTracker(Generic[_ProgressType]):
@@ -51,8 +50,8 @@ class ProgressTracker(Generic[_ProgressType]):
 
     async def set_progress(
         self,
-        state: Union[TaskState, str],
-        meta: Optional[_ProgressType] = None,
+        state: TaskState | str,
+        meta: _ProgressType | None = None,
     ) -> None:
         """Set progress.
 
@@ -73,7 +72,7 @@ class ProgressTracker(Generic[_ProgressType]):
             progress,
         )
 
-    async def get_progress(self) -> Optional[TaskProgress[_ProgressType]]:
+    async def get_progress(self) -> TaskProgress[_ProgressType] | None:
         """Get progress."""
         return await self.context.broker.result_backend.get_progress(
             self.context.message.task_id,
