@@ -2,7 +2,7 @@ import os
 from logging import getLogger
 from pathlib import Path
 from tempfile import gettempdir
-from typing import Any, Optional
+from typing import Any
 
 from taskiq.abc.middleware import TaskiqMiddleware
 from taskiq.message import TaskiqMessage
@@ -25,7 +25,7 @@ class PrometheusMiddleware(TaskiqMiddleware):
 
     def __init__(
         self,
-        metrics_path: Optional[Path] = None,
+        metrics_path: Path | None = None,
         server_port: int = 9000,
         server_addr: str = "0.0.0.0",  # noqa: S104
     ) -> None:
