@@ -1,12 +1,15 @@
 import asyncio
 from datetime import datetime, timezone
 from logging import getLogger
-from typing import Any, Optional
+from typing import Any
 from urllib.parse import urljoin
 
 import aiohttp
 
-from taskiq import TaskiqMessage, TaskiqMiddleware, TaskiqResult
+from taskiq.abc.middleware import TaskiqMiddleware
+from taskiq.compat import model_dump
+from taskiq.message import TaskiqMessage
+from taskiq.result import TaskiqResult
 
 __all__ = ("TaskiqAdminMiddleware",)
 
@@ -36,7 +39,7 @@ class TaskiqAdminMiddleware(TaskiqMiddleware):
         url: str,
         api_token: str,
         timeout: int = 5,
-        taskiq_broker_name: Optional[str] = None,
+        taskiq_broker_name: str | None = None,
     ) -> None:
         super().__init__()
         self.url = url
@@ -113,11 +116,13 @@ class TaskiqAdminMiddleware(TaskiqMiddleware):
 
         :param message: kicked message.
         """
+        dict_message: dict[str, Any] = model_dump(message)
         await self._spawn_request(
             f"/api/tasks/{message.task_id}/queued",
             {
-                "args": message.args,
-                "kwargs": message.kwargs,
+                "args": dict_message["args"],
+                "kwargs": dict_message["kwargs"],
+                "labels": dict_message["labels"],
                 "queuedAt": self._now_iso(),
                 "taskName": message.task_name,
                 "worker": self.__ta_broker_name,
@@ -134,11 +139,13 @@ class TaskiqAdminMiddleware(TaskiqMiddleware):
         :param message: incoming parsed taskiq message.
         :return: modified message.
         """
+        dict_message: dict[str, Any] = model_dump(message)
         await self._spawn_request(
             f"/api/tasks/{message.task_id}/started",
             {
-                "args": message.args,
-                "kwargs": message.kwargs,
+                "args": dict_message["args"],
+                "kwargs": dict_message["kwargs"],
+                "labels": dict_message["labels"],
                 "startedAt": self._now_iso(),
                 "taskName": message.task_name,
                 "worker": self.__ta_broker_name,
@@ -160,12 +167,13 @@ class TaskiqAdminMiddleware(TaskiqMiddleware):
         :param message: incoming message.
         :param result: result of execution for current task.
         """
+        dict_result: dict[str, Any] = model_dump(result)
         await self._spawn_request(
             f"/api/tasks/{message.task_id}/executed",
             {
                 "finishedAt": self._now_iso(),
                 "executionTime": result.execution_time,
                 "error": None if result.error is None else repr(result.error),
-                "returnValue": {"return_value": result.return_value},
+                "returnValue": {"return_value": dict_result["return_value"]},
             },
         )
