@@ -1,9 +1,8 @@
 import asyncio
-from datetime import datetime, timedelta
-from typing import Any, Dict, List
+from datetime import datetime, timedelta, timezone
+from typing import Any
 
 import pytest
-import pytz
 from freezegun import freeze_time
 
 from taskiq.brokers.inmemory_broker import InMemoryBroker
@@ -18,9 +17,9 @@ from taskiq.scheduler.scheduler import TaskiqScheduler
     "schedule_label",
     [
         pytest.param([{"cron": "* * * * *"}], id="cron"),
-        pytest.param([{"time": datetime.now(pytz.UTC)}], id="time"),
+        pytest.param([{"time": datetime.now(timezone.utc)}], id="time"),
         pytest.param(
-            [{"time": datetime.now(pytz.UTC), "labels": {"foo": "bar"}}],
+            [{"time": datetime.now(timezone.utc), "labels": {"foo": "bar"}}],
             id="labels_inside_schedule",
         ),
         pytest.param(
@@ -29,7 +28,7 @@ from taskiq.scheduler.scheduler import TaskiqScheduler
         ),
     ],
 )
-async def test_label_discovery(schedule_label: List[Dict[str, Any]]) -> None:
+async def test_label_discovery(schedule_label: list[dict[str, Any]]) -> None:
     broker = InMemoryBroker()
 
     @broker.task(
@@ -92,8 +91,8 @@ async def test_task_scheduled_at_time_runs_only_once(mock_sleep: None) -> None:
         @broker.task(
             task_name="test_task",
             schedule=[
-                {"time": datetime.now(pytz.UTC), "args": [1]},
-                {"time": datetime.now(pytz.UTC) + timedelta(days=1), "args": [2]},
+                {"time": datetime.now(timezone.utc), "args": [1]},
+                {"time": datetime.now(timezone.utc) + timedelta(days=1), "args": [2]},
                 {"cron": "1 * * * *", "args": [3]},
             ],
         )
