@@ -1,4 +1,4 @@
-from taskiq_redis import ListQueueBroker, RedisScheduleSource
+from taskiq_redis import ListQueueBroker, ListRedisScheduleSource
 
 from taskiq import TaskiqScheduler
 
@@ -6,7 +6,7 @@ from taskiq import TaskiqScheduler
 broker = ListQueueBroker("redis://localhost:6379/0")
 
 # Here's the source that is used to store scheduled tasks
-redis_source = RedisScheduleSource("redis://localhost:6379/0")
+redis_source = ListRedisScheduleSource("redis://localhost:6379/0")
 
 # And here's the scheduler that is used to query scheduled sources
 scheduler = TaskiqScheduler(broker, sources=[redis_source])
@@ -15,4 +15,4 @@ scheduler = TaskiqScheduler(broker, sources=[redis_source])
 @broker.task
 async def my_task(arg1: int, arg2: str) -> None:
     """Example task."""
-    print("Hello from my_task!", arg1, arg2)  # noqa: T201
+    print("Hello from my_task!", arg1, arg2)
