@@ -1,6 +1,7 @@
+from collections.abc import Callable
 from logging import getLogger
 from pathlib import Path
-from typing import Any, Callable
+from typing import Any
 
 from gitignore_parser import parse_gitignore
 from watchdog.events import FileSystemEvent
@@ -45,7 +46,7 @@ class FileWatcher:  # pragma: no cover
                 return
         except Exception as exc:
             logger.info(
-                f"Cannot check path `{event.src_path}` in gitignore. Cause: {exc}",
+                f"Cannot check path `{event.src_path!r}` in gitignore. Cause: {exc}",
             )
             return
 
