@@ -1,7 +1,8 @@
 import inspect
 import logging
+from collections.abc import Callable
 from dataclasses import dataclass
-from typing import Any, Callable, get_type_hints
+from typing import Any, get_type_hints
 
 import pytest
 from pydantic import BaseModel
