@@ -1,7 +1,7 @@
 from typing import Any
+from unittest.mock import AsyncMock
 
 import pytest
-from mock import AsyncMock
 
 from taskiq.exceptions import ResultIsReadyError, TaskiqResultTimeoutError
 from taskiq.funcs import gather
@@ -14,8 +14,8 @@ async def test_gather() -> None:
     rb_mock.is_result_ready.return_value = True
     rb_mock.get_result.return_value = 1
 
-    task1: "AsyncTaskiqTask[Any]" = AsyncTaskiqTask(task_id="1", result_backend=rb_mock)
-    task2: "AsyncTaskiqTask[Any]" = AsyncTaskiqTask(task_id="2", result_backend=rb_mock)
+    task1: AsyncTaskiqTask[Any] = AsyncTaskiqTask(task_id="1", result_backend=rb_mock)
+    task2: AsyncTaskiqTask[Any] = AsyncTaskiqTask(task_id="2", result_backend=rb_mock)
 
     assert await gather(task1, task2) == (1, 1)  # type: ignore
 
@@ -25,8 +25,8 @@ async def test_gather_timeout() -> None:
     rb_mock = AsyncMock()
     rb_mock.is_result_ready.return_value = False
 
-    task1: "AsyncTaskiqTask[Any]" = AsyncTaskiqTask(task_id="1", result_backend=rb_mock)
-    task2: "AsyncTaskiqTask[Any]" = AsyncTaskiqTask(task_id="2", result_backend=rb_mock)
+    task1: AsyncTaskiqTask[Any] = AsyncTaskiqTask(task_id="1", result_backend=rb_mock)
+    task2: AsyncTaskiqTask[Any] = AsyncTaskiqTask(task_id="2", result_backend=rb_mock)
 
     with pytest.raises(TaskiqResultTimeoutError):
         await gather(task1, task2, timeout=0.4)
@@ -37,8 +37,8 @@ async def test_gather_result_backend_error() -> None:
     rb_mock = AsyncMock()
     rb_mock.is_result_ready.side_effect = Exception
 
-    task1: "AsyncTaskiqTask[Any]" = AsyncTaskiqTask(task_id="1", result_backend=rb_mock)
-    task2: "AsyncTaskiqTask[Any]" = AsyncTaskiqTask(task_id="2", result_backend=rb_mock)
+    task1: AsyncTaskiqTask[Any] = AsyncTaskiqTask(task_id="1", result_backend=rb_mock)
+    task2: AsyncTaskiqTask[Any] = AsyncTaskiqTask(task_id="2", result_backend=rb_mock)
 
     with pytest.raises(ResultIsReadyError):
         await gather(task1, task2)
